+++ HISTORY.md
@@ -1,5 +1,18 @@
 # 1.1.0
   * Добавлено
+    - Класс `Environment`.
+    - Класс `Version`.
+    - Класс `StringBuilder`.
+    - Класс `DataTable` (`DataColumnCollection`, `DataRowCollection`, `DataColumn`, `DataRow`).
+    - Расширение `replaceAt` для класса `String`.
+    - Расширение `replaceAll` для класса `String`.
+    - Расширение `format` для класса `String`.
+    - Расширение `ranrom` для класса `String`.
+  * Дополнено
+    - Класс `Path`.
+
+# 1.1.0
+  * Добавлено
     - Класс `Color`.
     - Расширение `round` для класса `Number`.